Fixed Lorenz solution. Added Pendulum solution and graphs.
Changed the Lorenz solution in FinalProject to increase the number of
steps and changed the state variables to be initialized as floats. This
fixed the solution so the graph shows the two attractors and models
closer to Scipys odeint.
Started adding comparison of Scipys odeint with the Lorenz solution in
RungeKuttaTest.

Added dampened driven nonlinear pendulum function and modeled the
difference in the angles for two similiar starting conditions in a
semilogy plot.