Added pendulum equation derivation and a number of figures.
Added the derivation of the equation for the pendulum problem,
and a small introduction to explain the overall project.

Added graphs of the pendulum with 0.5 and 1.2 driving force values,
graphing the angular difference in two pendulums with an initial theta
difference of 0.001 radians. Also added approximate lines to show the
Lyapunov exponent for each case.

Added a number of figures to graph the Lorenz equations solutions to
a few different starting values.

Unsuccesfully added work for animations both through matplotlib's
animationFunction as well as turtle. No implementation would work.