Adds explanation of variables in Lorenz equ
Explains x, y, and z in the Lorenz equation to give
context of the problem.

Labels the subplots for the first solution to the
Lorenz equations to allow easier reference in the
description.

Increases the number of steps used in the calculation
of the first solution to the Lorenz equations for
better accuracy.